+++ R/csv_to_parquet.R
@@ -12,17 +12,10 @@
 #'
 #' }
 #'
-#' @param path_to_csv string that indicates the path or url to the csv file
-#' @param url_to_csv DEPRECATED use path_to_csv instead
-#' @param csv_as_a_zip DEPRECATED
 #' @param filename_in_zip name of the csv file in the zip. Required if several csv are included in the zip.
-#' @param path_to_parquet string that indicates the path to the directory where the parquet file will be stored
-#' @param columns character vector of columns to select from the input file (by default, all columns are selected).
-#' @param compression compression algorithm. Default "snappy".
-#' @param compression_level compression level. Meaning depends on compression algorithm.
-#' @param partition string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file.
-#' If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.
-#' @param encoding string that indicates the character encoding for the input file.
+#' @param url_to_csv DEPRECATED use path_to_file instead
+#' @param csv_as_a_zip DEPRECATED
+#' @inheritParams table_to_parquet
 #' @param ... additional format-specific arguments, see \href{https://arrow.apache.org/docs/r/reference/write_parquet.html}{arrow::write_parquet()}
 #'  and \href{https://arrow.apache.org/docs/r/reference/write_dataset.html}{arrow::write_dataset()} for more informations.
 #'
@@ -31,69 +24,51 @@
 #' see \href{https://rdrr.io/r/utils/unzip.html}{here}).
 #' In this case, it's advised to unzip your csv file by hand
 #' (for example with \href{https://www.7-zip.org/}{7-Zip})
-#' then use the function with the argument `path_to_csv`.
+#' then use the function with the argument `path_to_file`.
 #'
 #' @return A parquet file, invisibly
 #'
-#' @importFrom readr read_delim locale
-#' @importFrom curl curl_download
-#' @importFrom arrow write_parquet
-#' @importFrom utils unzip
-#' @importFrom cli cli_alert_danger cli_progress_message cli_alert_success cli_alert_info
-#' @importFrom tidyselect all_of everything
-#' @importFrom lifecycle deprecate_warn deprecated
 #' @export
 #'
 #' @examples
 #'
-#'
 #' # Conversion from a local csv file to a single parquet file :
 #'
 #' csv_to_parquet(
-#'   path_to_csv = parquetize_example("region_2022.csv"),
-#'   path_to_parquet = tempdir()
+#'   path_to_file = parquetize_example("region_2022.csv"),
+#'   path_to_parquet = tempfile(fileext=".parquet")
 #' )
 #'
 #' # Conversion from a local csv file to a single parquet file and select only
-#' # fex columns :
+#' # few columns :
 #'
 #' csv_to_parquet(
-#'   path_to_csv = parquetize_example("region_2022.csv"),
-#'   path_to_parquet = tempdir(),
+#'   path_to_file = parquetize_example("region_2022.csv"),
+#'   path_to_parquet = tempfile(fileext = ".parquet"),
 #'   columns = c("REG","LIBELLE")
 #' )
 #'
-#' # Conversion from a local csv file  to a partitioned parquet file  :
+#' # Conversion from a local csv file to a partitioned parquet file  :
 #'
 #' csv_to_parquet(
-#'   path_to_csv = parquetize_example("region_2022.csv"),
-#'   path_to_parquet = tempdir(),
+#'   path_to_file = parquetize_example("region_2022.csv"),
+#'   path_to_parquet = tempfile(fileext = ".parquet"),
 #'   partition = "yes",
 #'   partitioning =  c("REG")
-#' )
-#'
-#' # Conversion from a URL and a csv file with "gzip" compression :
-#'
-#' csv_to_parquet(
-#'   path_to_csv =
-#'   "https://github.com/sidsriv/Introduction-to-Data-Science-in-python/raw/master/census.csv",
-#'   path_to_parquet = tempdir(),
-#'   compression = "gzip",
-#'   compression_level = 5
 #' )
 #'
 #' # Conversion from a URL and a zipped file :
 #'
 #' csv_to_parquet(
-#'   path_to_csv = "https://www.nomisweb.co.uk/output/census/2021/census2021-ts007.zip",
+#'   path_to_file = "https://www.nomisweb.co.uk/output/census/2021/census2021-ts007.zip",
 #'   filename_in_zip = "census2021-ts007-ctry.csv",
-#'   path_to_parquet = tempdir()
+#'   path_to_parquet = tempfile(fileext = ".parquet")
 #' )
 
 csv_to_parquet <- function(
-    path_to_csv,
-    url_to_csv = deprecated(),
-    csv_as_a_zip = deprecated(),
+    path_to_file,
+    url_to_csv = lifecycle::deprecated(),
+    csv_as_a_zip = lifecycle::deprecated(),
     filename_in_zip,
     path_to_parquet,
     columns = "all",
@@ -107,7 +82,7 @@ csv_to_parquet <- function(
     lifecycle::deprecate_warn(
       when = "0.5.5",
       what = "csv_to_parquet(url_to_csv)",
-      details = "This argument is replaced by path_to_csv."
+      details = "This argument is replaced by path_to_file."
     )
   }
 
@@ -119,32 +94,28 @@ csv_to_parquet <- function(
     )
   }
 
-  # Check if at least one of the two arguments path_to_csv or url_to_csv is set
-  if (missing(path_to_csv) & missing(url_to_csv)) {
-    cli_alert_danger("Be careful, you have to fill in either the path_to_csv or url_to_csv argument")
-    stop("")
+  # Check if at least one of the two arguments path_to_file or url_to_csv is set
+  if (missing(path_to_file) & missing(url_to_csv)) {
+    cli_abort("Be careful, you have to fill the path_to_file argument", class = "parquetize_missing_argument")
   }
 
   # Check if path_to_parquet is missing
   if (missing(path_to_parquet)) {
-    cli_alert_danger("Be careful, the argument path_to_parquet must be filled in")
-    stop("")
+    cli_abort("Be careful, the argument path_to_parquet must be filled in", class = "parquetize_missing_argument")
   }
-
-  dir.create(path_to_parquet, recursive = TRUE, showWarnings = FALSE)
 
   # Check if columns argument is a character vector
   if (isFALSE(is.vector(columns) & is.character(columns))) {
-    cli_alert_danger("Be careful, the argument columns must be a character vector")
-    cli_alert_info('You can use `all` or `c("col1", "col2"))`')
-    stop("")
+    cli_abort(c("Be careful, the argument columns must be a character vector",
+              'You can use `all` or `c("col1", "col2"))`'),
+              class = "parquetize_bad_argument")
   }
 
-  if (missing(path_to_csv)) {
-    path_to_csv <- url_to_csv
+  if (missing(path_to_file)) {
+    path_to_file <- url_to_csv
   }
 
-  input_file <- download_extract(path_to_csv, filename_in_zip)
+  input_file <- download_extract(path_to_file, filename_in_zip)
 
   Sys.sleep(0.01)
   cli_progress_message("Reading data...")
@@ -157,14 +128,13 @@ csv_to_parquet <- function(
     col_select = if (identical(columns,"all")) everything() else all_of(columns)
   )
 
+  dataset <- write_parquet_at_once(
+    csv_output,
+    path_to_parquet,
+    partition,
+    compression,
+    compression_level,
+    ...)
 
-  Sys.sleep(0.01)
-  cli_progress_message("Writing data...")
-
-  parquetname <- get_parquet_file_name(input_file)
-  parquetfile <- write_data_in_parquet(csv_output, path_to_parquet, parquetname, partition, ...)
-
-  cli_alert_success("\nThe csv file is available in parquet format under {path_to_parquet}")
-
-  return(invisible(parquetfile))
+  return(invisible(dataset))
 }