+++ R/download_extract.R
@@ -17,7 +17,6 @@
 #'
 #' @return the path to the usable (uncompressed) file, invisibly.
 #'
-#' @importFrom tools file_ext
 #' @export
 #'
 #' @examples
@@ -28,7 +27,7 @@
 #' file_path <- download_extract(system.file("extdata","mtcars.csv.zip", package = "readr"))
 #' csv_to_parquet(
 #'   file_path,
-#'   path_to_parquet = tempdir()
+#'   path_to_parquet = tempfile(fileext = ".parquet")
 #' )
 #'
 #' # 1. download a remote file
@@ -41,7 +40,7 @@
 #' )
 #' csv_to_parquet(
 #'   file_path,
-#'   path_to_parquet = tempdir()
+#'   path_to_parquet = tempfile(fileext = ".parquet")
 #' )
 #'
 #' # the file is local and not zipped so :
@@ -50,7 +49,7 @@
 #' file_path <- download_extract(parquetize_example("region_2022.csv"))
 #' csv_to_parquet(
 #'   file_path,
-#'   path_to_parquet = tempdir()
+#'   path_to_parquet = tempfile(fileext = ".parquet")
 #' )
 #'
 download_extract <- function(path, filename_in_zip) {
@@ -66,8 +65,8 @@ download_extract <- function(path, filename_in_zip) {
   names(csv_files) <- basename(csv_files)
 
   if (length(csv_files) > 1 & missing(filename_in_zip)) {
-    cli_alert_danger("Be careful, zip files contains more than one file, you must set filename_in_zip argument")
-    stop("")
+    cli_abort("Be careful, zip files contains more than one file, you must set filename_in_zip argument",
+              class = "parquetize_missing_argument")
   } else if (length(csv_files) > 1) {
     path <- csv_files[[filename_in_zip]]
   } else {