+++ R/rds_to_parquet.R
@@ -13,16 +13,11 @@
 #'
 #' }
 #'
-#' @param path_to_rds string that indicates the path to the rds file
-#' @param path_to_parquet string that indicates the path to the directory where the parquet file will be stored
-#' @param partition string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file.
-#' If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.
+#' @inheritParams table_to_parquet
 #' @param ... additional format-specific arguments, see \href{https://arrow.apache.org/docs/r/reference/write_parquet.html}{arrow::write_parquet()}
 #'  and \href{https://arrow.apache.org/docs/r/reference/write_dataset.html}{arrow::write_dataset()} for more informations.
 #' @return A parquet file, invisibly
 #'
-#' @importFrom arrow write_parquet write_dataset
-#' @importFrom cli cli_alert_danger cli_progress_message cli_alert_success
 #' @export
 #'
 #' @examples
@@ -30,53 +25,51 @@
 #' # Conversion from a local rds file to a single parquet file ::
 #'
 #' rds_to_parquet(
-#'   path_to_rds = system.file("extdata","iris.rds",package = "parquetize"),
-#'   path_to_parquet = tempdir()
+#'   path_to_file = system.file("extdata","iris.rds",package = "parquetize"),
+#'   path_to_parquet = tempfile(fileext = ".parquet")
 #' )
 #'
 #' # Conversion from a local rds file to a partitioned parquet file  ::
 #'
 #' rds_to_parquet(
-#'   path_to_rds = system.file("extdata","iris.rds",package = "parquetize"),
-#'   path_to_parquet = tempdir(),
+#'   path_to_file = system.file("extdata","iris.rds",package = "parquetize"),
+#'   path_to_parquet = tempfile(fileext = ".parquet"),
 #'   partition = "yes",
 #'   partitioning =  c("Species")
 #' )
 
 rds_to_parquet <- function(
-    path_to_rds,
+    path_to_file,
     path_to_parquet,
     partition = "no",
+    compression = "snappy",
+    compression_level = NULL,
     ...
 ) {
 
-  # Check if path_to_rds is missing
-  if (missing(path_to_rds)) {
-    cli_alert_danger("Be careful, the argument path_to_rds must be filled in")
-    stop("")
+  # Check if path_to_file is missing
+  if (missing(path_to_file)) {
+    cli_abort("Be careful, the argument path_to_file must be filled in", class = "parquetize_missing_argument")
   }
 
   # Check if path_to_parquet is missing
   if (missing(path_to_parquet)) {
-    cli_alert_danger("Be careful, the argument path_to_parquet must be filled in")
-    stop("")
+    cli_abort("Be careful, the argument path_to_parquet must be filled in", class = "parquetize_missing_argument")
   }
-
-  dir.create(path_to_parquet, recursive = TRUE, showWarnings = FALSE)
 
   Sys.sleep(0.01)
   cli_progress_message("Reading data...")
 
-  rds_output <- readRDS(file = path_to_rds)
+  rds_output <- readRDS(file = path_to_file)
 
-  Sys.sleep(0.01)
-  cli_progress_message("Writing data...")
+  dataset <- write_parquet_at_once(
+    rds_output,
+    path_to_parquet,
+    partition,
+    compression,
+    compression_level,
+    ...)
 
-  parquetname <- get_parquet_file_name(path_to_rds)
-  parquetfile <- write_data_in_parquet(rds_output, path_to_parquet, parquetname, partition, ...)
-
-  cli_alert_success("\nThe rds file is available in parquet format under {path_to_parquet}")
-
-  return(invisible(parquetfile))
+  return(invisible(dataset))
 
 }